feat(app): split lightbox, add footer scroll driven
