+++ src/routes/Home/components/Breadcrumbs/index.tsx
@@ -5,12 +5,12 @@ const { Item } = BreadcrumbsNest;
 const Breadcrumbs = () => {
   return (
     <nav css={{ margin: '1rem 0 1.5rem' }}>
-      <BreadcrumbsNest overflow={4} css={{ marginBottom: '1rem' }}>
+      <BreadcrumbsNest overflow={5} css={{ marginBottom: '1rem' }}>
         <Item href="#">Home</Item>
-        <Item href="#">PC & Laptop</Item>
-        <Item href="#">Laptop</Item>
-        <Item href="#">Consumer Laptop</Item>
-        <Item>Apple Macbook Pro 2021 M1 Max 16 512GB</Item>
+        <Item href="#">Mainan & Hobi</Item>
+        <Item href="#">Model Kit</Item>
+        <Item href="#">Mecha Model (Gunpla)</Item>
+        <Item>Gundam Fix Figuration Metal Composite/GFF Metal Composite Wing Zero EW</Item>
       </BreadcrumbsNest>
     </nav>
   );
